Better fallback choosing in command search
Without this patch we have no way to execute the exit callback of
interface configuration in the clitest.c example!

So I've changed the command search algorithm to the following:

- If a command exists with our current mode setting, then
we choose that one.
- If our current mode is higher then MODE_CONFIG,
but a command with MODE_CONFIG exists, we choose that one, and change
our current mode down to MODE_CONFIG.
- If the first two method fails, we choose a command with MODE_ANY,
if exists.