Fix valgrind report issue in cli_loop() for long command entry
Valgrind reported the following error if a 'full' buffer of text read.
Can be demonstrated by setting CLI_MAX_LINE_LENGTH to 100 and entering
a long line.

==69152== Invalid read of size 1
==69152==    at 0x4C2CA44: __GI_strlen (vg_replace_strmem.c:459)
==69152==    by 0x50F97FD: strdup (strdup.c:41)
==69152==    by 0x40390A: cli_add_history (libcli.c:665)
==69152==    by 0x406B5B: cli_loop (libcli.c:1933)
==69152==    by 0x40225E: main (clitest.c:333)
==69152==  Address 0x5843d54 is 0 bytes after a block of size 100 alloc'd
==69152==    at 0x4C29C23: malloc (vg_replace_malloc.c:299)
==69152==    by 0x40524F: cli_loop (libcli.c:1213)
==69152==    by 0x40225E: main (clitest.c:333)

The issue is caused by no check being done to see if current cursor is
already at the 'last' character in the cmd buffer before appending the
new character to the cmd buffer.  This can result in a non-NULL
terminated string being passed to cli_add_history() and then the
strdup() walking off the end of the array.  Fixed by checking to see
if we are already at the end of the cmd buffer prior to appending a
character.  If not, add it, otherwise explicitly set the last
character to a NULL.