Fix valgrind reported issue in cli_clear_line() w/long lines
Valgrind reported the following error if a 'long' command is entered
and then 'ctrl-u' entered to clear the line.  Can be demonstrated by
setting CLI_MAX_LINE_LENGTH to 100 and and entering a line at least
34 characters long.

==11531== Invalid write of size 1
==11531==    at 0x404DBA: cli_clear_line (libcli.c:1118)
==11531==    by 0x405B8D: cli_loop (libcli.c:1541)
==11531==    by 0x4021C6: main (clitest.c:333)
==11531==  Address 0x5843d54 is 0 bytes after a block of size 100 alloc'd
==11531==    at 0x4C29C23: malloc (vg_replace_malloc.c:299)
==11531==    by 0x4050D6: cli_loop (libcli.c:1200)
==11531==    by 0x4021C6: main (clitest.c:333)
==11531==

The problem is that cli_clear_line() appends all of the characters
required to clear the cmd buffer into the cmd buffer itself and then
writes this out once.  If the number of characters in the buffer is
greater than 1/3 of the maximum buffer size cli_clear_line() will walk
off the end of the buffer.

Fixed by doing 3 separate writes (once to backspace to the beginning,
once to write spaces, and once to backspace back to the beginning
again).