Valgrind indirectly reported issue in cli_loop while interting text
Valgrind indirectly reported an issue while interting text into a full
command buffer.  Can be demonstrated by setting CLI_MAX_LINE_LENGTH to
100, entering a long line, then inserting a character anywhere except
at the end of the command and immediately deleting it.

==66078== Invalid read of size 1
==66078==    at 0x40599F: cli_loop (libcli.c:1501)
==66078==    by 0x4021C6: main (clitest.c:333)
==66078==  Address 0x5843d54 is 0 bytes after a block of size 100 alloc'd
==66078==    at 0x4C29C23: malloc (vg_replace_malloc.c:299)
==66078==    by 0x4050D6: cli_loop (libcli.c:1200)
==66078==    by 0x4021C6: main (clitest.c:333)
==66078==

While the error is reported in the backspace code, the cause is in the
insertion code.  There is no check to see if you walk off the end
of the cmd buffer while moving the charaters one space to the left.
If the initial buffer is 'full', the null terminator that was at the
end of the cmd buffer gets overwritten character immediately preceeding
it, and the 'end of buffer' index is set to CLI_MAX_LINE_LENGTH.  The
backspace code then proceeds to try and read beyond the end of the
buffer (moving cmd[CLI_MAX_LINE_LENGTH] to cmd[CLI_MAX_LINE_LENGTH-1].

Fixed by using memmove to handle moving the characters after the
insertion one place left, then checking to see if we are at the end
of the cmd buffer.  If so, we do not adjust 'l' (the end of buffer
index) and explicitly force cmd[l] to NULL to preserve a null
terminated string.  Otherwise we increment 'l'.