Fix possible double free in cli_command_name
Coverity indicates a possible double free in cli_command_name.  This
appears to be due to name being set to cli->commandname, and then if
this is not null name is freed, but cli->commandname is untouched.  If
this routine exits early (for example, the calloc fails), then
cli->commandname is now pointing to a freed memory block.

Fixed by ensuring that cli->commandname is freed and set to null if
not null already before assigning cli->commandname to name.