Fix valgrind reported issue in cli_loop for backspace w/long lines
Valgrind reported the following error when deleting a character from
anywhere except the very end.  Can be demonstrated by setting
CLI_MAX_LINE_LENGTH to 100 and entering a command of 100 characters,
then hitting 'left arrow' or ctrl-b and then 'delete'.

==35211== Invalid read of size 1
==35211==    at 0x40599F: cli_loop (libcli.c:1501)
==35211==    by 0x4021C6: main (clitest.c:333)
==35211==  Address 0x5843d54 is 0 bytes after a block of size 100 alloc'd
==35211==    at 0x4C29C23: malloc (vg_replace_malloc.c:299)
==35211==    by 0x4050D6: cli_loop (libcli.c:1200)
==35211==    by 0x4021C6: main (clitest.c:333)

The problem is that the loop 'moving' characters one space to the left
does not check to see it would walk off the end of the cmd buffer.

Fixed by using memmove handle the buffer movement, and explicitly
setting the former last character to null.  The modified line is
written out prior to actually making changes to the cmd buffer.