Fix varargs usage
Coverity found a number of 'missing varargs init or cleanup' issues in
the _print() method.  While examing these I found another potential
issue in the vasprintf() code, which is used only for WIN32 platforms.
Our target is Linux, and so that change is not tested.

The fix for _print() is to call vasprintf() directly without using a
loop.  The contents of cli->buffer (if any) will be freed before
updating cli->buffer and cli->buf_size with the new string results,
otherwise they are unchanged.  The fix for WIN32 only vasprintf() is
to wrap the first call to vsnprintf() with a va_copy()/va_end() to
preserve the passed in va_list.