Rework of command deletion to fix memory leak/segfault
There was at least one corner case in the delete a code pathway that
resulted in both a memory leak and traversing a no-longer valid pointer.
In 'cli_unregister_all()' cli->commands was never being updated to NULL.
Between 'cli_free_command()' and all callers there was no explicit
code to ensure that the any command pointing to the command being
deleted would point to the command after (if deleteing c, the c->next field).