Add 'CLI_CMD_SPOT_CHECK as a new optarg type
CLI_CMD_SPOT_CHECK is designed to allow a 'spot check' to ensure
that the optargs seen so far are 'sane' to allow parsing to continue.
The check will  *only* be done if there are no potential candidates
at the point where the spot check is being evaluated.  The recommended
place to put any spot check is immediately prior to any command that
would need 'sane' arguments to continue.