Major rework of how 'help' lines are formatted
All help lines consist of the name and helpstr.  This treats allows each
line to be 'wrapped' and collimated so each partial piece of helpstr will
be left aligned with each other.  Wrapping is done assuming an 80 col width,
and if a line is wrapped it will attempt to split it at the last white
space in that segment.  If none found, then it prints what it can.  Then it
also breaks at the first '\n' or '\r'.

Additionally, for optarg help, we can have 'specific' help messages in
addition to the general help for the optarg.  Look at the clitest program
for an example (both 'triangle' and 'rectangle' have specific help msgs
for the 'shape' optarg.  These specific help messages are tab ('\t')
separated and occur after the general help message.  A specific message
will only be shown if 'could' be a match.